+++ src/components/namespace-form/namespace-form.tsx
@@ -2,11 +2,9 @@ import * as React from 'react';
 import './namespace-form.scss';
 
 import { Form, FormGroup, TextInput, TextArea } from '@patternfly/react-core';
-
 import { MinusCircleIcon, PlusCircleIcon } from '@patternfly/react-icons';
 
-import { NamespaceCard } from '../cards/namespace-card';
-
+import { NamespaceCard } from '../../components';
 import { NamespaceType } from '../../api';
 
 interface IProps {
