+++ src/components/not-implemented/not-implemented.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+
 import { Link } from 'react-router-dom';
 
 export class NotImplemented extends React.Component<{}, {}> {
