+++ src/containers/partners/partners.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-import { withRouter, RouteComponentProps, Redirect } from 'react-router-dom';
+
+import { withRouter, RouteComponentProps } from 'react-router-dom';
+import { Main, Section } from '@redhat-cloud-services/frontend-components';
 
 import { BaseHeader, NotImplemented } from '../../components';
-
-import { Main, Section } from '@redhat-cloud-services/frontend-components';
 
 class Partners extends React.Component<RouteComponentProps, {}> {
     render() {
